Add tests
- Isolate tests
- Add CMakeLists.txt
