+++ tests/src/test_rvalues.cpp
@@ -1,0 +1,8 @@
+#include <dbg_macro/dbg.h>
+
+int main(int argc, char const* argv[]) {
+  std::string message = "hello world";
+  dbg(std::move(message));
+
+  return 0;
+}
